into binary -c: return 0 as single byte
The `into binary` command has a `-c` flag which strips any leading 0s
in the most significant digits to represent the minimal number of bytes,
rather than the system's complete in-memory representation of the input.

However, currently giving 0 as input results in eight 0 bytes even with
the `-c` flag, which is inconsistent with the purpose of the `-c` flag.

This change fixes this behavior so that if the entire input results in
all 0 bytes, only a single 0 byte is returned.