Use record api in `rotate`
TODO: Validate that the record assumptions are actually held true by
this command. There is the potential for invalid length as the
`old_column_names` seem to be taken from the last record (after
overwriting a bunch)

Furthermore we need to validate the column name uniqueness as well. This
command probably does not take care of that.