Refactor `split words`
Changed how the command works.  It now splits by ASCII whitespace.  Also
removed the length flags in favor of using the `filter` command.

This is marginally faster than the regex implementation and includes
digits in the words.

The main issue is that punctuation doesn't get removed or separated (so
`is.` and `is` are different).

Resolve #11168